Refactor the handlers out of connections
